+++ tests/integration/test_real_browser.py
@@ -51,12 +51,14 @@ class _ServerDescr:
 class IntegrationServers:
     """Integration servers starting/stopping manager"""
 
-    def __init__(self, *, loop=None):
+    def __init__(self, *, loop=None, use_resources=False):
         self.servers = {}
 
         self.loop = loop
         if self.loop is None:
             self.loop = asyncio.get_event_loop()
+
+        self.use_resources = use_resources
 
         self._logger = logging.getLogger("IntegrationServers")
 
@@ -178,7 +180,12 @@ class IntegrationServers:
             # doesn't support configuring for Resources.
             resource = server_descr.app.router["cors_resource"]
             route = next(iter(resource))
-            server_descr.cors.add(route)
+            if self.use_resources:
+                server_descr.cors.add(resource)
+                server_descr.cors.add(route)
+
+            else:
+                server_descr.cors.add(route)
 
     @asyncio.coroutine
     def stop_servers(self):
@@ -205,7 +212,8 @@ def _get_chrome_driver():
 class TestInBrowser(AioTestBase):
     @asyncio.coroutine
     def _test_in_webdriver(self, driver):
-        servers = IntegrationServers()
+        # TODO: Use pytest's fixtures to test use resources/not use resources.
+        servers = IntegrationServers(use_resources=True)
         yield from servers.start_servers()
 
         def selenium_thread():